splife

+++ events/guild/messageReactionAdd.js
@@ -64,12 +64,31 @@ module.exports = async (bot, reaction, user) => {
             var replyEmbed = new Discord.MessageEmbed()
                 .setColor(bot.config.COLORS.DENY)
                 .setDescription(`**Fermeture du ticket dans 10 secondes !**`)
+                .setFooter("Cliquez sur 🔓 pour réouvrir le ticket");
             let msg = await message.channel.send(replyEmbed);
+            msg.react("🔓");
             message.channel.ticketIsClosing = true;
             setTimeout(() => {
+                if(!message.channel.ticketIsClosing) return;
                 message.channel.delete()
             }, 10 * 1000)
            
+        }
+        if(reaction.emoji.name == "🔓"){
+            if(!message.channel.ticketIsClosing) return;
+            reaction.users.remove(user);
+            if(message.channel.ticketMember) message.channel.ticketMember.isInTicket = false;
+            if(!message.guild.members.cache.find(m => m.user.id == user.id).roles.cache.find(r => r.id == bot.config.I_ROLES.STAFF)) {
+                var replyEmbed = new Discord.MessageEmbed()
+                    .setColor(bot.config.COLORS.DENY)
+                    .setFooter(`Message auto-supprimé dans 5 secondes`)
+                    .setDescription(`<@${user.id}> **vous n'avez pas la permission de faire ca**`)
+                let msg = await message.channel.send(replyEmbed);
+                setTimeout(() => {msg.delete()}, 5 * 1000)
+                return;
+            }        
+            reaction.users.remove(bot.user);
+            message.channel.ticketIsClosing = false; 
         }
     }
 }
